% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{addOriginOfValue}
\alias{addOriginOfValue}
\title{Sets the OriginOfValues dataframe}
\usage{
addOriginOfValue(obj, index = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}}

\item{index}{A list of integer xxxxxxx}
}
\value{
An instance of class \code{MSnSet}.
}
\description{
Sets the OriginOfValues dataframe in the fData table
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
addOriginOfValue(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
