% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateIter}
\alias{aggregateIter}
\title{xxxx}
\usage{
aggregateIter(obj.pep, X, init.method = "Sum", method = "Mean", n = NULL)
}
\arguments{
\item{obj.pep}{xxxxx}

\item{X}{xxxx}

\item{init.method}{xxxxx}

\item{method}{xxxxx}

\item{n}{xxxx}
}
\value{
A protein object of class \code{MSnset}
}
\description{
Method to xxxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
protID <- "Protein_group_IDs"
X <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], protID, FALSE)
aggregateIter(Exp1_R25_pept[1:1000],X=X)
}
\author{
Samuel Wieczorek
}
