% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateIterParallel}
\alias{aggregateIterParallel}
\title{xxxx}
\usage{
aggregateIterParallel(
  obj.pep,
  X,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{obj.pep}{xxxxx}

\item{X}{xxxx}

\item{init.method}{xxxxx}

\item{method}{xxxxx}

\item{n}{xxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
protID <- "Protein_group_IDs"
obj.pep <- Exp1_R25_pept[1:1000]
X <- BuildAdjacencyMatrix(obj.pep, protID, FALSE)
aggregateIterParallel(obj.pep, X)
}
\author{
Samuel Wieczorek
}
