% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{boxPlotD}
\alias{boxPlotD}
\title{Builds a boxplot from a dataframe}
\usage{
boxPlotD(obj, conds, legend = NULL, palette = NULL)
}
\arguments{
\item{obj}{xxx}

\item{conds}{xxx}

\item{legend}{A vector of the conditions (one string per sample).}

\item{palette}{xxx}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
boxPlotD(Exp1_R25_pept, conds)
}
\seealso{
\code{\link{densityPlotD}}
}
\author{
Florence Combes, Samuel Wieczorek
}
