% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{boxPlotD_HC}
\alias{boxPlotD_HC}
\title{Builds a boxplot from a dataframe using the library \code{highcharter}}
\usage{
boxPlotD_HC(obj, legend = NULL, palette = NULL)
}
\arguments{
\item{obj}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data using the library \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
legend <- Biobase::pData(Exp1_R25_pept)[,"Sample.name"]
boxPlotD_HC(Exp1_R25_pept, legend)
}
\seealso{
\code{\link{densityPlotD_HC}}
}
\author{
Samuel Wieczorek
}
