% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{buildGraph}
\alias{buildGraph}
\title{Display a CC}
\usage{
buildGraph(The.CC, X)
}
\arguments{
\item{The.CC}{A cc (a list)}

\item{X}{xxxxx}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata') 
X <- BuildAdjacencyMatrix(Exp1_R25_pept, "Protein_group_IDs", FALSE)
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)
}
\author{
Thomas Burger, Samuel Wieczorek
}
