% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{check.conditions}
\alias{check.conditions}
\title{Check if the design is valid}
\usage{
\method{check}{conditions}(conds)
}
\arguments{
\item{conds}{A vector}
}
\value{
A list
}
\description{
This function check the validity of the conditions
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
check.conditions(Biobase::pData(Exp1_R25_pept)$Condition)
}
\author{
Samuel Wieczorek
}
