% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{check.design}
\alias{check.design}
\title{Check if the design is valid}
\usage{
\method{check}{design}(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the pData function of MSnbase}
}
\value{
A boolean
}
\description{
This function check the validity of the experimental design
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
check.design(Biobase::pData(Exp1_R25_pept)[,1:3])
}
\author{
Thomas Burger, Samuel Wieczorek
}
