% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{compareNormalizationD}
\alias{compareNormalizationD}
\title{Builds a plot from a dataframe}
\usage{
compareNormalizationD(
  qDataBefore,
  qDataAfter,
  condsForLegend = NULL,
  indData2Show = NULL,
  palette = NULL
)
}
\arguments{
\item{qDataBefore}{A dataframe that contains quantitative data before 
normalization.}

\item{qDataAfter}{A dataframe that contains quantitative data after 
normalization.}

\item{condsForLegend}{A vector of the conditions (one condition 
per sample).}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{palette}{xxx}
}
\value{
A plot
}
\description{
Plot to compare the quantitative proteomics data before and after 
normalization
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qDataBefore <- Biobase::exprs(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
objAfter <- wrapper.normalizeD(Exp1_R25_pept,"QuantileCentering","within conditions")
compareNormalizationD(qDataBefore, Biobase::exprs(objAfter), conds)
}
\author{
Samuel Wieczorek
}
