% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{compareNormalizationD_HC}
\alias{compareNormalizationD_HC}
\title{Builds a plot from a dataframe. Same as compareNormalizationD but 
uses the library \code{highcharter}}
\usage{
compareNormalizationD_HC(qDataBefore, qDataAfter, condsForLegend = NULL,
  indData2Show = NULL, palette = NULL)
}
\arguments{
\item{qDataBefore}{A dataframe that contains quantitative data before 
normalization.}

\item{qDataAfter}{A dataframe that contains quantitative data after 
normalization.}

\item{condsForLegend}{A vector of the conditions (one condition 
per sample).}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{palette}{xxx}
}
\value{
A plot
}
\description{
Plot to compare the quantitative proteomics data before and after 
normalization using the library \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:1000]
qDataBefore <- Biobase::exprs(obj)
conds <- Biobase::pData(obj)[,"Condition"]
objAfter <- wrapper.normalizeD(obj,"QuantileCentering","within conditions")
compareNormalizationD_HC(qDataBefore, Biobase::exprs(objAfter), conds)
}
\author{
Samuel Wieczorek
}
