% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{compareNormalizationDSubset}
\alias{compareNormalizationDSubset}
\title{Builds a plot from a dataframe}
\usage{
compareNormalizationDSubset(qDataBefore, qDataAfter, indData2Show = NULL,
  idsForLegend = NULL, palette = NULL, subset.view = NULL)
}
\arguments{
\item{qDataBefore}{A dataframe that contains quantitative data before 
normalization.}

\item{qDataAfter}{A dataframe that contains quantitative data after 
normalization.}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{idsForLegend}{A vector of the ids of the row in the data.frame qDataBefore.}

\item{palette}{xxx}

\item{subset.view}{A vector of index indicating rows to highlight}
}
\value{
A plot
}
\description{
Plot to compare the quantitative proteomics data before and after 
normalization for a subset of protein
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
qDataBefore <- Biobase::exprs(Exp1_R25_pept)
ids <-Exp1_R25_pept@featureData@data[,obj@experimentData@other$proteinId]
objAfter <- wrapper.normalizeD(Exp1_R25_pept,"QuantileCentering","within conditions")
compareNormalizationDSubset(qDataBefore, Biobase::exprs(objAfter), idsForLegend=ids,subset.view=1:10)
}
\author{
Samuel Wieczorek, Anais Courtier
}
