% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ttest.R
\name{compute.group.t.tests}
\alias{compute.group.t.tests}
\title{xxxxxx}
\usage{
compute.group.t.tests(
  qData,
  sTab,
  X,
  X.spec,
  logFC = NULL,
  contrast = "OnevsOne",
  type = "Student"
)
}
\arguments{
\item{qData}{A matrix of quantitative data, without any missing values.}

\item{sTab}{xxxx}

\item{X}{xxx}

\item{X.spec}{A matrix of quantitative data, without any missing values.}

\item{logFC}{A vector (or list of vectors) xxxx}

\item{contrast}{Indicates if the test consists of the comparison of each 
biological condition versus 
each of the other ones (contrast=1; 
for example H0:"C1=C2" vs H1:"C1!=C2", etc.) 
or each condition versus all others (contrast=2; e.g.  H0:"C1=(C2+C3)/2" vs
H1:"C1!=(C2+C3)/2", etc. if there are three conditions).}

\item{type}{xxxxx}
}
\value{
xxxxx
}
\description{
xxxxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept
protID <- "Protein_group_IDs"
keepThat <-  mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
X <- BuildAdjacencyMatrix(obj, protID,  unique = FALSE)
X.spec <- BuildAdjacencyMatrix(obj, protID,  unique = TRUE)
sTab <- Biobase::pData(obj)
qData <- Biobase::exprs(obj)
gttest <- compute.group.t.tests(qData, sTab, X, X.spec)
}
\author{
Thomas Burger, Samuel Wieczorek
}
