% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{createMSnset}
\alias{createMSnset}
\title{Creates an object of class \code{MSnSet} from text file}
\usage{
createMSnset(file, metadata = NULL, indExpData, indFData,
  indiceID = NULL, indexForOriginOfValue = NULL, logData = FALSE,
  replaceZeros = FALSE, pep_prot_data = NULL, proteinId = NULL,
  versions = NULL)
}
\arguments{
\item{file}{The name of a tab-separated file that contains the data.}

\item{metadata}{A dataframe describing the samples (in lines).}

\item{indExpData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{fData()} table
of the \code{MSnSet} object.}

\item{indFData}{The name of column in \code{file} that will be the name of
rows for the \code{exprs()} and \code{fData()} tables}

\item{indiceID}{The indice of the column containing the ID of entities 
(peptides or proteins)}

\item{indexForOriginOfValue}{xxxxxxxxxxx}

\item{logData}{A boolean value to indicate if the data have to be
log-transformed (Default is FALSE)}

\item{replaceZeros}{A boolean value to indicate if the 0 and NaN values of
intensity have to be replaced by NA (Default is FALSE)}

\item{pep_prot_data}{A string that indicates whether the dataset is about}

\item{proteinId}{xxxx}

\item{versions}{A list of the following items: Prostar_Version, DAPAR_Version
peptides or proteins.}
}
\value{
An instance of class \code{MSnSet}.
}
\description{
Builds an object of class \code{MSnSet} from a 
single tabulated-like file for quantitative and meta-data and a dataframe 
for the samples description. It differs from
the original \code{MSnSet} builder which requires three separated files 
tabulated-like quantitative proteomic data into a \code{MSnSet} object,
including metadata.
}
\examples{
require(DAPARdata)
require(Matrix)
exprsFile <- system.file("extdata", "Exp1_R25_pept.txt", package="DAPARdata")
metadataFile <- system.file("extdata", "samples_Exp1_R25.txt", package="DAPARdata")
metadata = read.table(metadataFile, header=TRUE, sep="\\t", as.is=TRUE)
indExpData <- c(56:61)
indFData <- c(1:55,62:71)
indiceID <- 64
createMSnset(exprsFile, metadata,indExpData,  indFData, indiceID, indexForOriginOfValue = NULL, pep_prot_data = "peptide")
}
\author{
Florence Combes, Samuel Wieczorek
}
