% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{CVDistD}
\alias{CVDistD}
\title{Distribution of CV of entities}
\usage{
CVDistD(qData, conds = NULL, palette = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}
}
\value{
A density plot
}
\description{
Builds a densityplot of the CV of entities in the exprs() table
of a object. The CV is calculated for each condition present
in the dataset (see the slot \code{'Condition'} in the \code{pData()} table)
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
CVDistD(Biobase::exprs(Exp1_R25_pept), conds)
}
\seealso{
\code{\link{densityPlotD}}.
}
\author{
Florence Combes, Samuel Wieczorek
}
