% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{deleteLinesFromIndices}
\alias{deleteLinesFromIndices}
\title{Delete the lines in the matrix of intensities and the metadata table
given their indice.}
\usage{
deleteLinesFromIndices(obj, deleteThat = NULL, processText = "")
}
\arguments{
\item{obj}{An object of class \code{MSnSet} containing
quantitative data.}

\item{deleteThat}{A vector of integers which are the indices of lines to 
delete.}

\item{processText}{A string to be included in the \code{MSnSet}
object for log.}
}
\value{
An instance of class \code{MSnSet} that have been filtered.
}
\description{
Delete the lines of \code{exprs()} table identified by their indice.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
deleteLinesFromIndices(Exp1_R25_pept, c(1:10))
}
\author{
Florence Combes, Samuel Wieczorek
}
