% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{densityPlotD}
\alias{densityPlotD}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD(obj, conds, legend = NULL, palette = NULL)
}
\arguments{
\item{obj}{xxx}

\item{conds}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
densityPlotD(Exp1_R25_pept, conds)
}
\seealso{
\code{\link{boxPlotD}}, \code{\link{CVDistD}}
}
\author{
Florence Combes, Samuel Wieczorek
}
