% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{densityPlotD_HC}
\alias{densityPlotD_HC}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD_HC(obj, legend = NULL, palette = NULL)
}
\arguments{
\item{obj}{xxx}

\item{legend}{A vector of the conditions (one condition 
per sample).}

\item{palette}{xxx}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples. Same as the function \code{\link{densityPlotD}}
but uses the package \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
densityPlotD_HC(Exp1_R25_pept)
}
\seealso{
\code{\link{boxPlotD}}, \code{\link{CVDistD}}
}
\author{
Samuel Wieczorek
}
