% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaGetSignificant}
\alias{diffAnaGetSignificant}
\title{Returns a MSnSet object with only proteins significant after 
differential analysis.}
\usage{
diffAnaGetSignificant(obj)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}
}
\value{
A MSnSet
}
\description{
Returns a MSnSet object with only proteins significant after 
differential analysis.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept
keepThat <- mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
qData <- Biobase::exprs(obj)
sTab <- Biobase::pData(obj)
allComp <- limmaCompleteTest(qData,sTab)
data <- list(logFC=allComp$logFC[1], P_Value = allComp$P_Value[1])
obj <- diffAnaSave(obj, allComp, data)
signif <- diffAnaGetSignificant(obj)
}
\author{
Alexia Dorffer
}
