% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaSave}
\alias{diffAnaSave}
\title{Returns a \code{MSnSet} object with the results of
the differential analysis performed with \code{\link{limma}} package.}
\usage{
diffAnaSave(obj, allComp, data = NULL, th_pval = 0, th_logFC = 0)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{allComp}{A list of two items which is the result of the function wrapper.limmaCompleteTest or xxxx}

\item{data}{The result of the differential analysis processed 
by \code{\link{limmaCompleteTest}}}

\item{th_pval}{xxx}

\item{th_logFC}{xxx}
}
\value{
A MSnSet
}
\description{
This method returns a class \code{MSnSet} object with the results
of differential analysis.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept
keepThat <- mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
qData <- Biobase::exprs(obj)
sTab <- Biobase::pData(obj)
allComp <- limmaCompleteTest(qData,sTab)
data <- list(logFC=allComp$logFC[1], P_Value = allComp$P_Value[1])
diffAnaSave(obj, allComp, data)
}
\author{
Alexia Dorffer, Samuel Wieczorek
}
