% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{display.CC.visNet}
\alias{display.CC.visNet}
\title{Display a CC}
\usage{
display.CC.visNet(g, layout = layout_nicely, obj = NULL,
  prot.tooltip = NULL, pept.tooltip = NULL)
}
\arguments{
\item{g}{A cc (a list)}

\item{layout}{xxxxx}

\item{obj}{xxx}

\item{prot.tooltip}{xxx}

\item{pept.tooltip}{xxx}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept) 
X <- BuildAdjacencyMatrix(Exp1_R25_pept, "Protein.group.IDs", FALSE)
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)
display.CC.visNet(g)
}
\author{
Thomas Burger, Samuel Wieczorek
}
