% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{findMECBlock}
\alias{findMECBlock}
\title{Finds the LAPALA into a \code{MSnSet} object}
\usage{
findMECBlock(obj)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}
}
\value{
A data.frame that contains the indexes of LAPALA
}
\description{
This method finds the LAPALA in a dataset.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:1000]
lapala <- findMECBlock(obj)
}
\author{
Samuel Wieczorek
}
