% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{get.pep.prot.cc}
\alias{get.pep.prot.cc}
\title{Build the list of connex composant of the adjacency matrix}
\usage{
get.pep.prot.cc(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of CC
}
\description{
Method to build the list of connex composant of the adjacency matrix
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata') 
X <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], "Protein_group_IDs", FALSE)
ll <- get.pep.prot.cc(X)
}
\author{
Thomas Burger, Samuel Wieczorek
}
