% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{Get_AllComparisons}
\alias{Get_AllComparisons}
\title{Returns list that contains a list of the statistical tests performed with DAPAR and recorded
in an object of class \code{MSnSet}.}
\usage{
Get_AllComparisons(obj)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}
}
\value{
A list of two slots: logFC and P_Value
}
\description{
This method returns a list of the statistical tests performed with DAPAR and recorded
in an object of class \code{MSnSet}.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept
keepThat <- mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
qData <- Biobase::exprs(obj)
sTab <- Biobase::pData(obj)
allComp <- limmaCompleteTest(qData,sTab)
data <- list(logFC=allComp$logFC[1], P_Value = allComp$P_Value[1])
obj <- diffAnaSave(obj, allComp, data)
ll <- Get_AllComparisons(obj)
}
\author{
Samuel Wieczorek
}
