% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{getIndicesOfLinesToRemove}
\alias{getIndicesOfLinesToRemove}
\title{Get the indices of the lines to delete, based on a prefix string}
\usage{
getIndicesOfLinesToRemove(obj, idLine2Delete = NULL, prefix = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{idLine2Delete}{The name of the column that correspond to the data 
to filter}

\item{prefix}{A character string that is the prefix to find in the data}
}
\value{
A vector of integers.
}
\description{
This function returns the indice of the lines to delete, based on a 
prefix string
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
getIndicesOfLinesToRemove(Exp1_R25_pept, "Potential.contaminant", prefix="+")
}
\author{
Samuel Wieczorek
}
