% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getListNbValuesInLines}
\alias{getListNbValuesInLines}
\title{Returns the possible number of values in lines in the data}
\usage{
getListNbValuesInLines(obj, type = "wholeMatrix")
}
\arguments{
\item{obj}{An object of class \code{MSnSet}}

\item{type}{xxxxxxx}
}
\value{
An integer
}
\description{
Returns the possible number of values in lines in a matrix.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
getListNbValuesInLines(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
