% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GetNbPeptidesUsed}
\alias{GetNbPeptidesUsed}
\title{Computes the number of peptides used for aggregating each protein}
\usage{
GetNbPeptidesUsed(X, pepData)
}
\arguments{
\item{X}{An adjacency matrix}

\item{pepData}{A data.frame of quantitative data}
}
\value{
A data.frame
}
\description{
Method to compute the number of quantified peptides used for aggregating each protein
}
\author{
Samuel Wieczorek
}
