% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{getNumberOf}
\alias{getNumberOf}
\title{Number of lines with prefix}
\usage{
getNumberOf(obj, name = NULL, prefix = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{name}{The name of a column.}

\item{prefix}{A string}
}
\value{
An integer
}
\description{
Returns the number of lines, in a given column, where content matches 
the prefix.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
getNumberOf(Exp1_R25_pept, "Potential_contaminant", "+")
}
\author{
Samuel Wieczorek
}
