% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{getQuantile4Imp}
\alias{getQuantile4Imp}
\title{Quantile imputation value definition}
\usage{
getQuantile4Imp(qData, qval = 0.025, factor = 1)
}
\arguments{
\item{qData}{An expression set containing quantitative values of various replicates}

\item{qval}{The quantile used to define the imputation value}

\item{factor}{A scaling factor to multiply the imputation value with}
}
\value{
A list of two vectors, respectively containing the imputation values and the rescaled imputation values
}
\description{
This method returns the q-th quantile of each colum of an expression set, up to a scaling factor
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qData <- Biobase::exprs(Exp1_R25_pept)
getQuantile4Imp(qData) 
}
\author{
Thomas Burger
}
