% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForGOAnalysis}
\alias{getTextForGOAnalysis}
\title{Build the text information for the Aggregation process}
\usage{
getTextForGOAnalysis(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Build the text information for the Aggregation process
}
\examples{
getTextForGOAnalysis(list())
}
\author{
Samuel Wieczorek
}
