% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForHypothesisTest}
\alias{getTextForHypothesisTest}
\title{Build the text information for the hypothesis test process}
\usage{
getTextForHypothesisTest(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Builds the text information for the hypothesis test process
}
\examples{
params <- list(design='OnevsOne', method='limma')
getTextForHypothesisTest(params)
}
\author{
Samuel Wieczorek
}
