% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForNewDataset}
\alias{getTextForNewDataset}
\title{Build the text information for a new dataset}
\usage{
getTextForNewDataset(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Build the text information for a new dataset
}
\examples{
getTextForNewDataset(list(filename="foo.MSnset"))
}
\author{
Samuel Wieczorek
}
