% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForNormalization}
\alias{getTextForNormalization}
\title{Build the text information for the Normalization process}
\usage{
getTextForNormalization(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Build the text information for the Normalization process
}
\examples{
getTextForNormalization(list(method="SumByColumns"))
}
\author{
Samuel Wieczorek
}
