% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logText.R
\name{getTextForpeptideImputation}
\alias{getTextForpeptideImputation}
\title{Build the text information for the peptide Imputation process}
\usage{
getTextForpeptideImputation(l.params)
}
\arguments{
\item{l.params}{A list of parameters related to the process of the dataset}
}
\value{
A string
}
\description{
Build the text information for the peptide Imputation process
}
\examples{
params <- list()
getTextForpeptideImputation(params)
}
\author{
Samuel Wieczorek
}
