% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GraphPepProt}
\alias{GraphPepProt}
\title{Function to create a histogram that shows the repartition of
peptides w.r.t. the proteins}
\usage{
GraphPepProt(mat)
}
\arguments{
\item{mat}{An adjacency matrix.}
}
\value{
A histogram
}
\description{
Method to create a plot with proteins and peptides on
a MSnSet object (peptides)
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
mat <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], "Protein.group.IDs")
GraphPepProt(mat)
}
\author{
Alexia Dorffer, Samuel Wieczorek
}
