% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ttest.R
\name{groupttest}
\alias{groupttest}
\title{xxxxxx}
\usage{
groupttest(MatAdj, cond1 = NULL, cond2 = NULL)
}
\arguments{
\item{MatAdj}{xxxxx.}

\item{cond1}{xxxx.}

\item{cond2}{xxxx.}
}
\value{
xxxxx
}
\description{
This function is xxxxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept
protID <- "Protein_group_IDs"
keepThat <-  mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
X.spec <- BuildAdjacencyMatrix(obj, protID,  unique = TRUE)
qData <- Biobase::exprs(obj)
gttest <- groupttest(X.spec, qData[,1:3], qData[,4:6])
}
\author{
Thomas Burger, Samuel Wieczorek
}
