% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{hc_logFC_DensityPlot}
\alias{hc_logFC_DensityPlot}
\title{Density plots of logFC values}
\usage{
hc_logFC_DensityPlot(df_logFC, threshold_LogFC = 0, palette = NULL)
}
\arguments{
\item{df_logFC}{A dataframe that contains the logFC values}

\item{threshold_LogFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}

\item{palette}{xxx}
}
\value{
A highcharts density plot
}
\description{
This function show the density plots of Fold Change (the same as calculated by limma) for a list 
of the comparisons of conditions in a differnetial analysis.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept[1:1000]
keepThat <- mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
qData <- Biobase::exprs(obj)
sTab <- Biobase::pData(obj)
res <- limmaCompleteTest(qData,sTab)
hc_logFC_DensityPlot(res$logFC)
}
\author{
Samuel Wieczorek
}
