% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{heatmap_HC}
\alias{heatmap_HC}
\title{This function is inspired from the function \code{\link{heatmap.2}} 
that displays quantitative data in the \code{exprs()} table of an object of
class \code{MSnSet}. For more information, please refer to the help 
of the heatmap.2 function.}
\usage{
heatmap_HC(qData, col = heat.colors(100), labCol)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{col}{colors used for the image. Defaults to heat colors (heat.colors).}

\item{labCol}{character vectors with column conds to use.}
}
\value{
A heatmap
}
\description{
Heatmap inspired by the heatmap.2 function and uses the highcharts library
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept[1:100,]
obj <- mvFilter(obj, "wholeMatrix", 6)
qData <- Biobase::exprs(obj)
heatmap_HC(qData)
}
\author{
Samuel Wieczorek
}
