% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.iter}
\alias{inner.aggregate.iter}
\title{xxxx}
\usage{
inner.aggregate.iter(pepData, X, init.method = "Sum", method = "Mean",
  n = NULL)
}
\arguments{
\item{pepData}{xxxxx}

\item{X}{xxxx}

\item{init.method}{xxx}

\item{method}{xxx}

\item{n}{xxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
require(DAPARdata)
data(Exp1_R25_pept)
protID <- "Protein.group.IDs"
obj.pep <- Exp1_R25_pept[1:1000]
X <- BuildAdjacencyMatrix(obj.pep, protID, FALSE)
DAPAR::inner.aggregate.iter(exprs(obj.pep), X)
}
