% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.topn}
\alias{inner.aggregate.topn}
\title{xxxx}
\usage{
inner.aggregate.topn(pepData, X, method = "Mean", n = 10)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}

\item{method}{xxxxx}

\item{n}{xxxxx}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
}
