% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.mean}
\alias{inner.mean}
\title{xxxx}
\usage{
inner.mean(pepData, X)
}
\arguments{
\item{pepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
xxxxx
}
\description{
Method to xxxxx
}
\author{
Samuel Wieczorek
}
