% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.MV}
\alias{is.MV}
\title{Similar to the function \code{is.na} but focused on the equality with the missing 
values in the dataset (type 'POV' and 'MEC')}
\usage{
is.MV(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na} but focused on the equality with the missing 
values in the dataset (type 'POV' and 'MEC')
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept
data <- Biobase::fData(obj)[,obj@experimentData@other$OriginOfValues]
is.MV(data)
}
\author{
Samuel Wieczorek
}
