% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepa.R
\name{LH1.lm}
\alias{LH1.lm}
\title{xxxxxx}
\usage{
LH1.lm(X, y1, y2, j)
}
\arguments{
\item{X}{an n.pep*n.prot indicator matrix.}

\item{y1}{n.pep*n.samples matrix giving the observed counts for}

\item{y2}{n.pep*n.samples matrix giving the observed counts for}

\item{j}{the index of the protein being tested, ie which has different}
}
\value{
xxxxxxxxxx..
}
\description{
This function is xxxxxxx
}
\author{
Thomas Burger, Laurent Jacob
}
