% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{listSheets}
\alias{listSheets}
\title{This function returns the list of the sheets names in a Excel file.}
\usage{
listSheets(file)
}
\arguments{
\item{file}{The name of the Excel file.}
}
\value{
A vector
}
\description{
This function lists all the sheets of an Excel file.
}
\author{
Samuel Wieczorek
}
