% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{make.design.3}
\alias{make.design.3}
\title{Builds the design matrix for designs of level 3}
\usage{
make.design.3(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the pData function of MSnbase}
}
\value{
A design matrix
}
\description{
This function builds the design matrix for design of level 3
}
\examples{
\donttest{
require(DAPARdata)
data(Exp1_R25_pept)
make.design.3(Biobase::pData(Exp1_R25_pept))
}
}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
