% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{mvFilter}
\alias{mvFilter}
\title{Filter lines in the matrix of intensities w.r.t. some criteria}
\usage{
mvFilter(obj, type, th, processText = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet} containing
quantitative data.}

\item{type}{Method used to choose the lines to delete.
Values are : "None", "wholeMatrix", "allCond", "atLeastOneCond"}

\item{th}{An integer value of the threshold}

\item{processText}{A string to be included in the \code{MSnSet}
object for log.}
}
\value{
An instance of class \code{MSnSet} that have been filtered.
}
\description{
Filters the lines of \code{exprs()} table with conditions on the number
of missing values.
The user chooses the minimum amount of intensities that is acceptable and
the filter delete lines that do not respect this condition.
The condition may be on the whole line or condition by condition.
}
\details{
The different methods are :
"wholeMatrix": given a threshold \code{th}, only the lines that contain
at least \code{th} values are kept.
"allCond": given a threshold \code{th}, only the lines which contain
at least \code{th} values for each of the conditions are kept.
"atLeastOneCond": given a threshold \code{th}, only the lines that contain
at least \code{th} values, and for at least one condition, are kept.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
mvFilter(Exp1_R25_pept, "wholeMatrix", 2)
}
\author{
Florence Combes, Samuel Wieczorek
}
