% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvHisto}
\alias{mvHisto}
\title{Histogram of missing values}
\usage{
mvHisto(
  qData,
  samplesData,
  conds,
  indLegend = "auto",
  showValues = FALSE,
  palette = NULL
)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and 
columns to the meta-data for those samples.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{indLegend}{The indices of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}

\item{palette}{xxx}
}
\value{
A histogram
}
\description{
This method plots a histogram of missing values.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qData <- Biobase::exprs(Exp1_R25_pept)
samplesData <- Biobase::pData(Exp1_R25_pept)
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
mvHisto(qData, samplesData, conds, indLegend="auto", showValues=TRUE)
}
\author{
Florence Combes, Samuel Wieczorek
}
