% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvPerLinesHisto}
\alias{mvPerLinesHisto}
\title{Bar plot of missing values per lines}
\usage{
mvPerLinesHisto(qData, samplesData, indLegend = "auto", showValues = FALSE)
}
\arguments{
\item{qData}{A dataframe that contains the data to plot.}

\item{samplesData}{A dataframe which contains informations about 
the replicates.}

\item{indLegend}{The indice of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A bar plot
}
\description{
This method plots a bar plot which represents the distribution of the 
number of missing values (NA) per lines (ie proteins).
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qData <- Biobase::exprs(Exp1_R25_pept)
samplesData <- Biobase::pData(Exp1_R25_pept)
mvPerLinesHisto(qData, samplesData)
}
\author{
Florence Combes, Samuel Wieczorek
}
