% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{mvPerLinesHistoPerCondition}
\alias{mvPerLinesHistoPerCondition}
\title{Bar plot of missing values per lines and per condition}
\usage{
mvPerLinesHistoPerCondition(
  qData,
  samplesData,
  indLegend = "auto",
  showValues = FALSE,
  palette = NULL
)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{samplesData}{A dataframe where lines correspond to samples and 
columns to the meta-data for those samples.}

\item{indLegend}{The indice of the column name's in \code{pData()} tab}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}

\item{palette}{xxx}
}
\value{
A bar plot
}
\description{
This method plots a bar plot which represents the distribution of the 
number of missing values (NA) per lines (ie proteins) and per conditions.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
qData <- Biobase::exprs(Exp1_R25_pept)
samplesData <- Biobase::pData(Exp1_R25_pept)
mvPerLinesHistoPerCondition(qData, samplesData)
}
\author{
Samuel Wieczorek
}
