% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_hc_chart}
\alias{my_hc_chart}
\title{Customised resetZoomButton of highcharts plots}
\usage{
my_hc_chart(hc, chartType, zoomType = "None")
}
\arguments{
\item{hc}{A highcharter object}

\item{chartType}{The type of the plot}

\item{zoomType}{The type of the zoom (one of "x", "y", "xy", "None")}
}
\value{
A highchart plot
}
\description{
Customise the resetZoomButton of highcharts plots.
}
\examples{
library("highcharter")
hc <- highchart() 
hc_chart(hc,type = "line") 
hc_add_series(hc,data = c(29, 71, 40))
my_hc_ExportMenu(hc,filename='foo')
}
\author{
Samuel Wieczorek
}
