% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_hc_ExportMenu}
\alias{my_hc_ExportMenu}
\title{Customised contextual menu of highcharts plots}
\usage{
my_hc_ExportMenu(hc, filename)
}
\arguments{
\item{hc}{A highcharter object}

\item{filename}{The filename under which the plot has to be saved}
}
\value{
A contextual menu for highcharts plots
}
\description{
Customise the contextual menu of highcharts plots.
}
\examples{
library("highcharter")
hc <- highchart() 
hc_chart(hc,type = "line") 
hc_add_series(hc,data = c(29, 71, 40))
my_hc_ExportMenu(hc,filename='foo')
}
\author{
Samuel Wieczorek
}
