% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{NumericalgetIndicesOfLinesToRemove}
\alias{NumericalgetIndicesOfLinesToRemove}
\title{Get the indices of the lines to delete, based on a prefix string}
\usage{
NumericalgetIndicesOfLinesToRemove(obj, name = NULL, value = NULL,
  operator = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{name}{The name of the column that correspond to the data to filter}

\item{value}{xxxx}

\item{operator}{A xxxx}
}
\value{
A vector of integers.
}
\description{
This function returns the indice of the lines to delete, based on a 
prefix string
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
NumericalgetIndicesOfLinesToRemove(Exp1_R25_pept, "A.Count", value="6", operator='==')
}
\author{
Samuel Wieczorek
}
