% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{plotJitter_rCharts}
\alias{plotJitter_rCharts}
\title{Display a a jitter plot for CC}
\usage{
plotJitter_rCharts(df, clickFunction = NULL)
}
\arguments{
\item{df}{xxxx}

\item{clickFunction}{xxxx}
}
\value{
A plot
}
\description{
Display a jitter plot for CC
}
\examples{
\dontrun{
}
}
\author{
Thomas Burger, Samuel Wieczorek
}
