% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{plotPCA_Eigen}
\alias{plotPCA_Eigen}
\title{Plots the eigen values of PCA}
\usage{
plotPCA_Eigen(res.pca)
}
\arguments{
\item{res.pca}{xxx}
}
\value{
A histogram
}
\description{
Plots the eigen values of PCA
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
res.pca <- wrapper.pca(Exp1_R25_pept, ncp=6)
plotPCA_Eigen(res.pca)
}
\author{
Samuel Wieczorek
}
