% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{proportionConRev_HC}
\alias{proportionConRev_HC}
\title{Barplot of proportion of contaminants and reverse}
\usage{
proportionConRev_HC(nBoth = 0, nCont = 0, nRev = 0, lDataset = 0)
}
\arguments{
\item{nBoth}{The number of both contaminants and reverse identified in the dataset.}

\item{nCont}{The number of contaminants identified in the dataset.}

\item{nRev}{The number of reverse entities identified in the dataset.}

\item{lDataset}{The total length (number of rows) of the dataset}
}
\value{
A barplot
}
\description{
Plots a barplot of proportion of contaminants and reverse. Same as the function
\code{proportionConRev} but uses the package \code{highcharter}
}
\examples{
proportionConRev_HC(10, 20, 100)
}
\author{
Samuel Wieczorek
}
