% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prostar.R
\name{Prostar}
\alias{Prostar}
\title{Prostar}
\usage{
Prostar()
}
\value{
A new window in the default internet browser
}
\description{
Prostar : a GUI for DAPAR
}
\examples{
if(interactive()) {Prostar()}
}
\author{
Samuel Wieczorek
}

