% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{readExcel}
\alias{readExcel}
\title{This function reads a sheet of an Excel file and put the data into a data.frame.}
\usage{
readExcel(file, extension, sheet)
}
\arguments{
\item{file}{The name of the Excel file.}

\item{extension}{The extension of the file}

\item{sheet}{The name of the sheet}
}
\value{
A data.frame
}
\description{
This function reads a sheet of an Excel file and put the data into a data.frame.
}
\author{
Samuel Wieczorek
}
