% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesFilter.R
\name{removeLines}
\alias{removeLines}
\title{Removes lines in the dataset based on a prefix string.}
\usage{
removeLines(obj, idLine2Delete = NULL, prefix = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{idLine2Delete}{The name of the column that correspond to the 
data to filter}

\item{prefix}{A character string that is the prefix to find in the data}
}
\value{
An object of class \code{MSnSet}.
}
\description{
This function removes lines in the dataset based on a prefix string.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
removeLines(Exp1_R25_pept, "Potential_contaminant")
removeLines(Exp1_R25_pept, "Reverse")
}
\author{
Samuel Wieczorek
}
