% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{rep_row}
\alias{rep_row}
\title{Same as rep function but on a matrix.}
\usage{
rep_row(x, n)
}
\arguments{
\item{x}{The data to repeat}

\item{n}{The number of repetitions.}
}
\value{
A matrix
}
\description{
fill a matrix by row
}
\examples{
rep_row('foo', 4)
}
\author{
Samuel Wieczorek
}
