% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{saveParameters}
\alias{saveParameters}
\title{Saves the parameters of a tool in the pipeline of Prostar}
\usage{
saveParameters(obj, name.dataset = NULL, name = NULL,
  l.params = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}}

\item{name.dataset}{The name of the dataset}

\item{name}{The name of the tool. Available values are: "Norm, Imputation, anaDiff, GOAnalysis,Aggregation"}

\item{l.params}{A list that contains the parameters}
}
\value{
An instance of class \code{MSnSet}.
}
\description{
Saves the parameters of a tool in the pipeline of Prostar
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
l.params=list(method="Global quantile alignment", type="overall")
saveParameters(Exp1_R25_pept, "Filtered.peptide", "Imputation",l.params)
}
\author{
Samuel Wieczorek
}
