% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{setMEC}
\alias{setMEC}
\title{Sets the MEC tag in the OriginOfValues}
\usage{
setMEC(obj)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}}
}
\value{
An instance of class \code{MSnSet}.
}
\description{
Sets the MEC tag in the OriginOfValues
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
setMEC(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
