% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{splitAdjacencyMat}
\alias{splitAdjacencyMat}
\title{splits an adjacency matrix into specific and shared}
\usage{
splitAdjacencyMat(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of two adjacency matrices
}
\description{
Method to split an adjacency matrix into specific and shared
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata') 
X <- BuildAdjacencyMatrix(Exp1_R25_pept[1:1000], "Protein_group_IDs", FALSE)
X.ll <- splitAdjacencyMat(X)
}
\author{
Samuel Wieczorek
}
