% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{test.design}
\alias{test.design}
\title{Check if xxxxxx}
\usage{
test.design(tab)
}
\arguments{
\item{tab}{A data.frame which correspond to xxxxxx}
}
\value{
A list of two items
}
\description{
This function check xxxxx
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
test.design(Biobase::pData(Exp1_R25_pept)[,1:3])
}
\author{
Thomas Burger, Samuel Wieczorek
}
