% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_PeptideLevel.R
\name{translatedRandomBeta}
\alias{translatedRandomBeta}
\title{Generator of simulated values}
\usage{
translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)
}
\arguments{
\item{n}{An integer which is the number of simulation (same as in rbeta)}

\item{min}{An integer that corresponds to the lower bound of the interval}

\item{max}{An integer that corresponds to the upper bound of the interval}

\item{param1}{An integer that is the first parameter of rbeta function.}

\item{param2}{An integer that is second parameter of rbeta function.}
}
\value{
A vector of n simulated values
}
\description{
Generator of simulated values
}
\examples{
translatedRandomBeta(1000, 5, 10, 1, 1)
}
\author{
Thomas Burger
}
