% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{violinPlotD}
\alias{violinPlotD}
\title{Builds a violinplot from a dataframe}
\usage{
violinPlotD(obj, legend = NULL, palette = NULL, subset.view = NULL)
}
\arguments{
\item{obj}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{A vector of index indicating rows to highlight}
}
\value{
A violinplot
}
\description{
ViolinPlot for quantitative proteomics data
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
library(vioplot)
legend <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
violinPlotD(Exp1_R25_pept, legend=legend,subset.view=20:30)
}
\seealso{
\code{\link{densityPlotD}}
}
\author{
Samuel Wieczorek, Anais Courtier
}
