% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.compareNormalizationD_HC}
\alias{wrapper.compareNormalizationD_HC}
\title{Builds a plot from a dataframe}
\usage{
wrapper.compareNormalizationD_HC(
  objBefore,
  objAfter,
  condsForLegend = NULL,
  indData2Show = NULL,
  ...
)
}
\arguments{
\item{objBefore}{A dataframe that contains quantitative data before 
normalization.}

\item{objAfter}{A dataframe that contains quantitative data after 
normalization.}

\item{condsForLegend}{A vector of the conditions (one condition 
per sample).}

\item{indData2Show}{A vector of the indices of the columns to show in the 
plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{...}{arguments for palette}
}
\value{
A plot
}
\description{
Wrapper to the function that plot to compare the quantitative proteomics 
data before and after normalization. Same as the function \link{wrapper.compareNormalizationD}
but uses the package \code{highcharter}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
conds <- Biobase::pData(Exp1_R25_pept)[,"Condition"]
objAfter <- wrapper.normalizeD(Exp1_R25_pept, "QuantileCentering", 
"within conditions")
wrapper.compareNormalizationD_HC(Exp1_R25_pept, objAfter, conds)
}
\author{
Samuel Wieczorek
}
