% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.corrMatrixD}
\alias{wrapper.corrMatrixD}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table}
\usage{
wrapper.corrMatrixD(obj, rate = 5)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{rate}{A float that defines the gradient of colors.}
}
\value{
A colored correlation matrix
}
\description{
Builds a correlation matrix based on a \code{MSnSet} object.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.corrMatrixD(Exp1_R25_pept)
}
\author{
Alexia Dorffer
}
