% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.corrMatrixD_HC}
\alias{wrapper.corrMatrixD_HC}
\title{Displays a correlation matrix of the quantitative data of the
\code{exprs()} table}
\usage{
wrapper.corrMatrixD_HC(obj, rate = 0.5)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{rate}{A float that defines the gradient of colors.}
}
\value{
A colored correlation matrix
}
\description{
Builds a correlation matrix based on a \code{MSnSet} object. 
Same as the function \code{\link{wrapper.corrMatrixD}} but uses the package \code{highcharter}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.corrMatrixD_HC(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
