% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStats.R
\name{wrapper.CVDistD}
\alias{wrapper.CVDistD}
\title{Distribution of CV of entities}
\usage{
wrapper.CVDistD(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{...}{arguments for palette}
}
\value{
A density plot
}
\description{
Builds a densityplot of the CV of entities in the exprs() table
of an object \code{MSnSet}. The variance is calculated for each 
condition present
in the dataset (see the slot \code{'Condition'} in the \code{pData()} table).
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.CVDistD(Exp1_R25_pept)
}
\author{
Alexia Dorffer
}
