% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{wrapper.hc_mvTypePlot2}
\alias{wrapper.hc_mvTypePlot2}
\title{Distribution of observed values with respect to intensity values 
from a \code{MSnSet} object}
\usage{
wrapper.hc_mvTypePlot2(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{...}{See \code{\link{hc_mvTypePlot2}}}
}
\value{
A scatter plot
}
\description{
This method is a wrapper for the function \code{\link{hc_mvTypePlot2}} adapted to objects
of class \code{MSnSet}).
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
wrapper.hc_mvTypePlot2(Exp1_R25_pept)
}
\author{
Florence Combes, Samuel Wieczorek
}
