% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{wrapper.impute.detQuant}
\alias{wrapper.impute.detQuant}
\title{Wrapper of the function \code{\link{impute.detQuant}} for objects
of class \code{MSnSet}}
\usage{
wrapper.impute.detQuant(obj, qval = 0.025, factor = 1)
}
\arguments{
\item{obj}{An instance of class \code{MSnSet}}

\item{qval}{An expression set containing quantitative values of various replicates}

\item{factor}{A scaling factor to multiply the imputation value with}
}
\value{
An imputed instance of class \code{MSnSet}
}
\description{
This method is a wrapper of the function \code{\link{impute.detQuant}} for objects
of class \code{MSnSet}
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.impute.detQuant(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
