% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{wrapper.impute.fixedValue}
\alias{wrapper.impute.fixedValue}
\title{Missing values imputation from a \code{MSnSet} object}
\usage{
wrapper.impute.fixedValue(obj, fixVal)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{fixVal}{A float .}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This method is a wrapper to
objects of class \code{MSnSet} and imputes missing values with a fixed value.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.impute.fixedValue(Exp1_R25_pept[1:1000], 0.001)
}
\author{
Samuel Wieczorek
}
