% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{wrapper.impute.KNN}
\alias{wrapper.impute.KNN}
\title{KNN missing values imputation from a \code{MSnSet} object}
\usage{
wrapper.impute.KNN(obj, K)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{K}{the number of neighbors.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This method is a wrapper for
objects of class \code{MSnSet} and imputes missing values with a fixed value.
This function imputes the missing values condition by condition.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.impute.KNN(Exp1_R25_pept[1:1000], 3)
}
\author{
Samuel Wieczorek
}
