% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{wrapper.impute.pa}
\alias{wrapper.impute.pa}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
wrapper.impute.pa(obj, q.min = 0.025)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{q.min}{Same as the function \code{impute.pa} in the package \code{imp4p}}
}
\value{
The \code{exprs(obj)} matrix with imputed values instead of missing values.
}
\description{
This method is a wrapper to the function \code{impute.pa} of the package
\code{imp4p} adapted to an object of class \code{MSnSet}.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
dat <- mvFilter(Exp1_R25_pept[1:1000], type="allCond", th = 1)
dat <- wrapper.impute.pa(dat)
}
\author{
Samuel Wieczorek
}
