% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{wrapper.mvHisto_HC}
\alias{wrapper.mvHisto_HC}
\title{Histogram of missing values from a \code{MSnSet} object}
\usage{
wrapper.mvHisto_HC(obj, indLegend = "auto", showValues = FALSE, ...)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{indLegend}{The indices of the column name's in \code{pData()} tab.}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}

\item{...}{xxx}
}
\value{
A histogram
}
\description{
This method plots from a \code{MSnSet} object a histogram of 
missing values.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
wrapper.mvHisto_HC(Exp1_R25_pept, showValues=TRUE)
}
\author{
Alexia Dorffer
}
