% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{wrapper.mvPerLinesHisto_HC}
\alias{wrapper.mvPerLinesHisto_HC}
\title{Histogram of missing values per lines from an object using highcharter
\code{MSnSet}}
\usage{
wrapper.mvPerLinesHisto_HC(obj, indLegend = "auto", showValues = FALSE)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{indLegend}{The indice of the column name's in \code{pData()} tab .}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A histogram
}
\description{
This method is a wrapper to plots from a \code{MSnSet} object a 
histogram which represents the distribution of the 
number of missing values (NA) per lines (ie proteins).
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
wrapper.mvPerLinesHisto(Exp1_R25_pept)
}
\author{
Alexia Dorffer
}
