% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesPlots.R
\name{wrapper.mvPerLinesHistoPerCondition}
\alias{wrapper.mvPerLinesHistoPerCondition}
\title{Bar plot of missing values per lines and per conditions from an 
object \code{MSnSet}}
\usage{
wrapper.mvPerLinesHistoPerCondition(obj, indLegend = "auto",
  showValues = FALSE)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{indLegend}{The indice of the column name's in \code{pData()} tab .}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}
}
\value{
A bar plot
}
\description{
This method is a wrapper to plots from a \code{MSnSet} object a 
bar plot which represents the distribution of the 
number of missing values (NA) per lines (ie proteins) and per conditions.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.mvPerLinesHistoPerCondition(Exp1_R25_pept)
}
\author{
Samuel Wieczorek
}
