% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{wrapper.normalizeD}
\alias{wrapper.normalizeD}
\title{Normalisation}
\usage{
wrapper.normalizeD(obj, method, type = NULL, scaling = FALSE,
  quantile = 0.15, span = 0.7, subset.norm = NULL)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{method}{One of the following : "GlobalQuantileAlignment" (for
normalizations of important magnitude), "SumByColumns", "QuantileCentering",
"Mean Centering", "LOESS" and "vsn".}

\item{type}{For the method "Global Alignment", the parameters are:
"sum by columns": operates on the original scale (not the log2 one) and propose
to normalize each abundance by the total abundance of the sample (so as to focus
on the analyte proportions among each sample).
"Alignment on all quantiles": proposes to align the quantiles of all the
replicates; practically it amounts to replace abundances by order statistics.
For the two other methods, the parameters are "overall" (shift all the
sample distributions at once) or "within conditions" (shift the sample
distributions within each condition at a time).}

\item{scaling}{A boolean that indicates if the variance of the data have to
be forced to unit (variance reduction) or not.}

\item{quantile}{A float that corresponds to the quantile used to align the
data.}

\item{span}{parameter for LOESS method}

\item{subset.norm}{A vector of index indicating rows to be used for normalization}
}
\value{
An instance of class \code{MSnSet} where the quantitative
data in the \code{exprs()} tab has been normalized.
}
\description{
Provides several methods to normalize quantitative data from
a \code{MSnSet} object.
They are organized in six main families : GlobalQuantileAlignement, sumByColumns, QuantileCentering, MeanCentering, LOESS, vsn
For the first family, there is no type.
For the five other families, two type categories are available :
"Overall" which means that the value for each protein
(ie line in the expression data tab) is computed over all the samples ;
"within conditions" which means that the value for each protein
(ie line in the \code{exprs()} data tab) is computed condition
by condition.
Normalization can be based only on a subset of protein. Proteins with NA in this subset are ignored
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
wrapper.normalizeD(Exp1_R25_pept[1:1000], "QuantileCentering", "within conditions",subset.norm=1:10)
}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier
}
