% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{writeMSnsetToCSV}
\alias{writeMSnsetToCSV}
\title{Exports a MSnset dataset into a zip archive containing three zipped CSV files.}
\usage{
writeMSnsetToCSV(obj, fname)
}
\arguments{
\item{obj}{An object of class \code{MSnSet}.}

\item{fname}{The name of the archive file.}
}
\value{
A compressed file
}
\description{
This function exports a MSnset dataset into three csv files compressed in a zip file
}
\examples{
\donttest{
require(DAPARdata)
data(Exp1_R2_pept)
obj <- Exp1_R2_pept[1:1000]
writeMSnsetToCSV(obj, "foo")
}
}
\author{
Samuel Wieczorek
}
