% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggregateMean}
\alias{aggregateMean}
\title{Compute the intensity of proteins as the mean of the intensities
of their peptides.}
\usage{
aggregateMean(obj.pep, X)
}
\arguments{
\item{obj.pep}{A peptide object of class \code{MSnset}}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the mean of the 
intensities of their peptides.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj.pep <- Exp1_R25_pept[1:1000]
protID <- "Protein.group.IDs"
X <- BuildAdjacencyMatrix(obj.pep, protID, FALSE)
aggregateMean(obj.pep, X)
}
\author{
Alexia Dorffer
}
