% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{histPValue_HC}
\alias{histPValue_HC}
\title{Distribution of p-values (histogram)}
\usage{
histPValue_HC(pval_ll, bins = 80, pi0 = 1)
}
\arguments{
\item{pval_ll}{A list that contains the data}

\item{bins}{An integer that is the number of bins of the histogram.}

\item{pi0}{A float that is the threshold}
}
\value{
A plot
}
\description{
This function plots a distribution of the p-values.
}
\examples{
require(DAPARdata)
data(Exp1_R25_pept)
obj <- Exp1_R25_pept
keepThat <- mvFilterGetIndices(obj, 'wholeMatrix', ncol(obj))
obj <- mvFilterFromIndices(obj, keepThat)
qData <- Biobase::exprs(obj)
sTab <- Biobase::pData(obj)
allComp <- limmaCompleteTest(qData,sTab)
histPValue_HC(allComp$P_Value[1])
}
\author{
Samuel Wieczorek
}
