% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{splitAdjacencyMat}
\alias{splitAdjacencyMat}
\title{splits an adjacency matrix into specific and shared}
\usage{
splitAdjacencyMat(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of two adjacency matrices
}
\description{
Method to split an adjacency matrix into specific and shared
}
\author{
Samuel Wieczorek
}
